moves translatable file to i18n
